# API/Livro-openapi.yaml
openapi: "3.1.0"
info:
  title: "LibraryProject API"
  description: "LibraryProject API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
tags:
  - name: "Biblioteca INFNET - LivroController"
paths:
  /api/v1/livros/{id}:
    post:
      summary: "POST api/v1/livros/{id}"
      operationId: "criarLivro"
      parameters:
        - name: "id"
          in: "path"
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Livro"
        required: true
      responses:
        "200":
          description: "Livro criado com sucesso"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Livro"
        "400":
          description: "Requisição inválida - ID do livro não encontrada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Erro interno do servidor"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: "PUT api/v1/livros/{id}"
      operationId: "atualizarLivro"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Livro"
        required: true
      responses:
        "200":
          description: "Livro atualizado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Livro"
        "400":
          description: "Requisição inválida - ID de Livro não encontrada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Erro interno do servidor"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: "DELETE api/v1/livros/{id}"
      operationId: "deletarLivro"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "Livro deletado com sucesso"
        "400":
          description: "Requisição inválida - ID de livro não encontrada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Erro interno do servidor"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/livros:
    get:
      summary: "GET api/v1/livros"
      operationId: "listarLivros"
      responses:
        "200":
          description: "Livros recuperados com sucesso"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Livro"
        "204":
          description: "Nenhum conteúdo - Nenhum livro encontrado"
        "500":
          description: "Erro interno do servidor"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/livros/autor/{autor}:
    get:
      summary: "GET api/v1/livros/autor/{autor}"
      operationId: "listarLivrosPorAutor"
      parameters:
        - name: "autor"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "Livro encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Livro"
        "204":
          description: "Nenhum conteúdo - Livro não encontrado"
        "400":
          description: "Requisição inválida - Autor inválido"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Erro interno do servidor"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/livros/genero/{genero}:
    get:
      summary: "GET api/v1/livros/genero/{genero}"
      operationId: "listarLivrosPorGenero"
      parameters:
        - name: "genero"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "Livro encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Livro"
        "204":
          description: "Nenhum conteúdo - Livro não encontrado"
        "400":
          description: "Requisição inválida - Gênero inválido"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Erro interno do servidor"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/livros/id/{id}:
    get:
      summary: "GET api/v1/livros/id/{id}"
      operationId: "obterLivro"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "Livro encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Livro"
        "204":
          description: "Nenhum conteúdo - Livro não encontrado"
        "400":
          description: "Requisição inválida - ID do livro inválida"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Erro interno do servidor"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/livros/titulo/{titulo}:
    get:
      summary: "GET api/v1/livros/titulo/{titulo}"
      operationId: "listarLivrosPorTitulo"
      parameters:
        - name: "titulo"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "Livro encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Livro"
        "204":
          description: "Nenhum conteúdo - Livro não encontrado"
        "400":
          description: "Requisição inválida - Título inválido"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Erro interno do servidor"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Usuario:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        nome:
          type: "string"
        email:
          type: "string"
        senha:
          type: "string"
        meusLivros:
          type: "array"
          items:
            $ref: "#/components/schemas/Livro"
    Emprestimo:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        ativo:
          type: "boolean"
        livro:
          $ref: "#/components/schemas/Livro"
        usuario:
          $ref: "#/components/schemas/Usuario"
        dataVencimento:
          type: "string"
          format: "date-time"
    Livro:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        titulo:
          type: "string"
        autor:
          type: "string"
        genero:
          type: "string"
        isbn:
          type: "string"
        sinopse:
          type: "string"
        usuario:
          $ref: "#/components/schemas/Usuario"
        emprestimo:
          $ref: "#/components/schemas/Emprestimo"
    HttpHeaders:
      type: "object"
      properties: { }
    ResponseEntity:
      type: "object"
      properties:
        headers:
          $ref: "#/components/schemas/HttpHeaders"
      required:
        - "headers"
    Error:
      type: "object"
      properties:
        code:
          type: "integer"
        message: